Fix #9: Cracking the audio CAPTCHA is trivial
We harden the audio CAPTCHAs by dynamically adding white noise, so the
bit pattern is somewhat different for each CAPTCHA. While that's still
not very sophisticated, it appears to be good enough for our purposes.

To eschew the need for an MP3 codec, which is almost certainly not
available on shared hosting, we switch to outputting WAV, which isn't
as well supported as MP3, but users with unsupported browsers still can
download the file for playing. For input, we choose  mono *unsigned*
16-bit LPCM samples with a sampling rate of 8000 Hz in little-endian
byte order, which are simple to process (signed 16-bit LE is not
necessarily supported by unpack()).