Close #10: Don't litter the session
It appears to be acceptable to only store the most recent unsolved
CAPTCHA in the session, and to reuse it for subsequent calls to
`cryptographp_captcha_display()`. That may sometimes lead to a wrong
CAPTCHA being reported, but that is better than to litter the session
with potentially unnecessary information.