Improve visual CAPTCHA tests
Relying on comparisons of the hashes of the image files is way too
brittle; even small differences in the file structure would trigger
test failures, and we would not even have a chance to assess these
failures.

Instead we introduce a simple image comparison, which compares pre-
generated images to the images generated by Cryptographp_XH, and on
failure stores the latter plus simplistic diff images for manual
inspection.

We also drop the too simplistic `mt_rand()` mock, and seed the RNG
instead.  This is somewhat brittle, but makes the tests more suitable;
and being able to compare expected and actual output, as well as diffs
is supposed to make that bearable.