Set up scheduled workflow
Since we're relying on Plib_XH, we also need to catch potential
regressions regarding that.  Thus we set up a scheduled workflow.

To make that reasonably work, we must no longer push relevant changes
to the default branch after a release, since we want that release to be
tested against latest Plib_XH.  Instead we need to work in a feature
branch for the new release.

There is the issue, though, that scheduled workflows are automatically
disabled after 60 days of repository inactivity.  We need to figure out
how to handle this.