Improve error handling of CodeStore
So far we only put in some assertions to satisfy static analyzers.
However, this failures may actually happen in practice, so we convert
that to exceptions, which we catch in the public methods of `CodeStore`
which then signals failure.

The error handling is neither complete, nor particularly meaningful,
but we can still improve later.